:wrench: [REFACTOR] Logger 삭제
