+++ src/main/java/com/prochord/server/config/WebSocketConfig.java
@@ -1,14 +1,10 @@
 package com.prochord.server.config;
 
-import com.prochord.server.utils.WebSocketAuthInterceptor;
-import com.prochord.server.utils.WebSocketChatHandler;
+//import com.prochord.server.utils.WebSocketAuthInterceptor;
+import com.prochord.server.service.WebSocketChatHandler;
 import lombok.RequiredArgsConstructor;
-import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.*;
-
-import java.util.logging.Logger;
 
 @Configuration
 @EnableWebSocket
@@ -16,7 +12,7 @@ import java.util.logging.Logger;
 public class WebSocketConfig implements WebSocketConfigurer {
 
     private final WebSocketChatHandler webSocketChatHandler;
-    private final WebSocketAuthInterceptor webSocketAuthInterceptor;
+//    private final WebSocketAuthInterceptor webSocketAuthInterceptor;
 
 
 //    @Override
@@ -31,6 +27,8 @@ public class WebSocketConfig implements WebSocketConfigurer {
     @Override
     public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
         // WebSocket 연결할 엔드포인트 설정
-        registry.addHandler(webSocketChatHandler, "/chats").addInterceptors(webSocketAuthInterceptor).setAllowedOrigins("*");
+        registry.addHandler(webSocketChatHandler, "/chats")
+//                .addInterceptors(webSocketAuthInterceptor)
+                .setAllowedOrigins("*");
     }
 }
