+++ src/main/java/com/prochord/server/utils/WebSocketAuthInterceptor.java
@@ -1,49 +1,49 @@
-package com.prochord.server.utils;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.server.ServerHttpRequest;
-import org.springframework.http.server.ServerHttpResponse;
-import org.springframework.http.server.ServletServerHttpRequest;
-import org.springframework.http.server.ServletServerHttpResponse;
-import org.springframework.stereotype.Component;
-import org.springframework.web.socket.WebSocketHandler;
-import org.springframework.web.socket.server.HandshakeInterceptor;
-
-import java.util.Map;
-
-@Component
-@RequiredArgsConstructor
-public class WebSocketAuthInterceptor implements HandshakeInterceptor {
-
-    /**
-     * 웹소켓 연결 전 인터셉터
-     * Authorization 헤더를 확인하여 유저를 인증한다.
-     * 유저 명이 채팅유저로 시작하지 않으면 401을 반환한다.
-     * 유저 명이 채팅유저로 시작해서 인증된 유저라면 session에 username을 저장한다.
-     */
-    @Override
-    public boolean beforeHandshake(ServerHttpRequest request, ServerHttpResponse response,
-                                   WebSocketHandler wsHandler, Map<String, Object> attributes) throws Exception {
-        if (request instanceof ServletServerHttpRequest) {
-            HttpServletRequest servletRequest = ((ServletServerHttpRequest) request).getServletRequest();
-            HttpServletResponse servletResponse = ((ServletServerHttpResponse) response).getServletResponse();
-
-            String username = servletRequest.getHeader("Authorization");
-            if (username != null && username.startsWith("chatUser")) {
-                attributes.put("username", username);
-
-                return true;
-            } else {
-                servletResponse.setStatus(401);
-                return false;
-            }
-        }
-        return false;
-    }
-    @Override
-    public void afterHandshake(ServerHttpRequest request, ServerHttpResponse response,
-                               WebSocketHandler wsHandler, Exception exception) {
-    }
-}
+//package com.prochord.server.utils;
+//
+//import jakarta.servlet.http.HttpServletRequest;
+//import jakarta.servlet.http.HttpServletResponse;
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.http.server.ServerHttpRequest;
+//import org.springframework.http.server.ServerHttpResponse;
+//import org.springframework.http.server.ServletServerHttpRequest;
+//import org.springframework.http.server.ServletServerHttpResponse;
+//import org.springframework.stereotype.Component;
+//import org.springframework.web.socket.WebSocketHandler;
+//import org.springframework.web.socket.server.HandshakeInterceptor;
+//
+//import java.util.Map;
+//
+//@Component
+//@RequiredArgsConstructor
+//public class WebSocketAuthInterceptor implements HandshakeInterceptor {
+//
+//    /**
+//     * 웹소켓 연결 전 인터셉터
+//     * Authorization 헤더를 확인하여 유저를 인증한다.
+//     * 유저 명이 채팅유저로 시작하지 않으면 401을 반환한다.
+//     * 유저 명이 채팅유저로 시작해서 인증된 유저라면 session에 username을 저장한다.
+//     */
+//    @Override
+//    public boolean beforeHandshake(ServerHttpRequest request, ServerHttpResponse response,
+//                                   WebSocketHandler wsHandler, Map<String, Object> attributes) throws Exception {
+//        if (request instanceof ServletServerHttpRequest) {
+//            HttpServletRequest servletRequest = ((ServletServerHttpRequest) request).getServletRequest();
+//            HttpServletResponse servletResponse = ((ServletServerHttpResponse) response).getServletResponse();
+//
+//            String username = servletRequest.getHeader("Authorization");
+//            if (username != null && username.startsWith("chatUser")) {
+//                attributes.put("username", username);
+//
+//                return true;
+//            } else {
+//                servletResponse.setStatus(401);
+//                return false;
+//            }
+//        }
+//        return false;
+//    }
+//    @Override
+//    public void afterHandshake(ServerHttpRequest request, ServerHttpResponse response,
+//                               WebSocketHandler wsHandler, Exception exception) {
+//    }
+//}
